the pricing component is composed
